fix uptermRuntimeDir when /run/user/<uid> does not exist
After #398 was merged, `uptermRuntimeDir()` started using
uses `xdg.RuntimeDir` as base directory. When `XDG_RUNTIME_DIR`
is not set, this falls back to `/run/user/<uid>`.
In some non-interactive scenarios, `/run/user/<uid>` will not be
present and `upterm` won't be able to create it with
`os.MkdirAll` because the user does not have permissions to write to
`/run/user`, causing `upterm host` to fail.